obexd: Add workaround for devices which use absolute path in pbap Name header
The pbap spec says:

"5.1.2 Name
The Name header shall contain the absolute path in the virtual folders
architecture of the PSE, appended with the name of the file representation of
one of the Phone Book Objects.
Example: telecom/pb.vcf or SIM1/telecom/pb.vcf for the main phone book
objects."

The example actually uses relative paths but text state absolute paths although
the OBEX specification says that the Name header is always relative to the
current path.