Fix response on adapter RequestSession method
Fixes response on adapter RequestSession method to be sent after mode is
changed, if such is necessary. More specifically, change of power off
mode to power on is in question.

Currently response is sent when mode change is confirmed by agent, not
by response from controller. Such may lead to failed CreateDevice method
if it is called quickly enough after RequestSession when controller is
in powered off state.

New session is not created if there is already a session for such D-Bus
message.