Fix discovery callback usage
The start_discovery function was used both as a GLib callback as well as
directly called function that can return a negative integer as error.
This will obviously not work since GLib expects a gboolean return (a
negative return value would have an undefined result). This patch fixes
the issue creating a separate function to be used as the GLib callback.