Remove flags2type() function
This function was only being used by create_device_internal(). It is
generating incorrect information when a single mode LE device has flags
0x02 (General Discoverable Mode), it was being "detected" as Dual Mode,
and thus the adapter was attempting to connect through BR/EDR.

This commit fixes CreateDevice() for these devices, but it is a partial
solution. In future, we should drop the "device type" enumeration, as
this information will come from kernel "device found" events.

Also note that the "DEVICE_TYPE_DUALMODE" enumeration is not really
being used, so it will probably disappear too.