obexd: Fix contacts data cleanup for new missed calls
It is possible that phonebook_pull_read is invoked several times
submitting multiple pull requests without closing PBAP object. E.g.,
when history is large enough and maxlistcount>VCARDS_PART_COUNT.

The result is possibility of different data structures (GString and
contact_data) to be mixed in a single GSlist that may lead to undefined
behaviour.