main: Use glib signal handler
Writing anything to syslog in the signal hander is in danger
to deadlock with any ongoing syslog write (syslock is taking
a mutex). Fortunatly, BlueZ's signal handler doesn't do this,
therefore the change isn't strickly necessary, but the same changes
were done to oFono and ConnMan. So this patch streamlines BlueZ
with the other projects.