btmgmt: Add keys command
This command calls the load_link_keys mgmt method and can also be used
for load_long_term_keys in the future. Right now it always sends an
empty list but is still useful for activating the HCI_LINKKEYS flag in
the kernel so that the kernel keeps track of link keys.