Fix reading of on_mode in btd_adapter_get_mode
If InitiallyPowered was set to false, it was not possible to use
hciconfig tool, as on_mode variable was set to MODE_OFF making it
impossible to enable BT via hciconfig until a D-Bus call was done and the
variable was set in config file as connectable.

Additionally the patch removes the if-else condition, as the both calls
done exactly the same. Removed also last else condition as MODE_OFF
should not be valid for on_mode. Fix tested with hciops interface.