thermometer: Fix bad read operation when time stamp is not provided
Time stamp value is an optional field provided in the measure, so next
value in the array will start in a different index depending if the time
stamp was provided or not. This patch check this case and update the
index to a proper value in the byte array before reading the temperature
type value.