attrib-server: Attach attrib_channel in adapter clients list
This patch attaches attribute channels to the gatt server list instead
of using the global client list. Furthermore, we use the GATT server
started in the default adapter when attrib_channel_detach is invoked in
order to keep backward compatibility until the interface is updated to
accept the adapter too.