mgmt-api: Update to latest plans
- Remove the remote_name event and add the name to the device_connected
event (since we always refresh the name at that point anyway).

- Add Class of Device to device_connected since that's the only point
where we get it when a previously unknown device connects to us for the
first time.

- Add a Flags field to device_found to accommodate for legacy pairing
information (which we get from the remote host features HCI event) and
merge the only other existing boolean parameter (confirm_name) into this
Flags field.