core: Remove name and class setting from btd_adapter_start()
Especially for mgmtops it's important that these values are pushed down
in the stack *before* powering on, so btd_adapter_start() is the wrong
place. Instead, adapter functions are added to hciops/mgmtops can fetch
these values at the right point during adapter init and push them down
in the stack (to the kernel).