input: Set up uinput device again, if it was previously closed.
If you connect a PS3 controller to bluetoothd as an input device, then
take the batteries out of a PS3 controller, then put them back in and
push a few buttons, it will eventually cause an error which causes the
uinput socket to be closed. It will then re-connect to bluetoothd, but
the uinput socket fd will be -1, so it needs to be set up again after
the re-connect.