csr_bcsp: fix link establishment
The current link-establishment implementation times out if a packet is
not received after the link is active (i.e. reception of CONF_RESP).

Fix by making sure that the link establishment loop in open is
terminated when ubcsp sets the UBCSP_PACKET_SENT flag (on reception of
CONF_RESP) rather than UBCSP_PACKET_RECEIVED (on reception of any later
packages).