udev: remove udev activation of the blutooth daemon
Udev is not meant to spawn long-running processes, and as of the next
udev version such processes will be killed.

If using systemd, device activation can be used to get the same effect.

On Arch we have been disabling this rule for a long time, even though we
don't use systemd, as it was causing more problems than it was worth.