SAP: Don't append MaxMsgSize parameter to CONNECT_RESP unconditionaly
According to SAP 1.1 specification MaxMsgSize parameter of CONNECT_RESP
is conditional and should not be added for every status code. Add this
parameter only if status code is  "Error, Server does not support
message size" (see section 4.1.1 of SAP specification).

To avoid confusion maxmsgsize param is removed from sap_connect_rsp and
it is assumed that SAP driver is able to handle messages with size up
to SAP_BUF_SIZE bytes.