SAP: Make sap_error_rsp function static
This error message is used to reply to SAP protocol errors and it
should not be used by drivers as they don't have access to server
internal state machine. Messages passed to drivers are valid and
properly formatted so drivers shall use result code in response
messages to indicate error (section 4.11 of SAP specification).