attrib-server: Fix MTU for GATT
This patch changes the size of the input and output buffers to the same
size of the channel MTU. The channel MTU, in turn, is the L2CAP MTU for
BR/EDR channels, and the ATT MTU for LE channels. This value can change
between calls to channel_handler if a GATT MTU exchange procedure is
performed, which is only supported by LE channels.